Move grouping logic from api to data models
